+++ sdk/src/object_crawler/wrappers.rs
@@ -51,37 +51,134 @@ where
     }
 }
 
+fn new_from_tag_and_fields_id<'de, D, T>(
+    deserializer: D,
+    constructor: impl FnOnce(sui::MoveTypeTag, sui::UID) -> T,
+) -> Result<T, D::Error>
+where
+    D: Deserializer<'de>,
+{
+    #[derive(Deserialize)]
+    struct Wrapper {
+        #[serde(rename = "type")]
+        type_: String,
+        fields: ObjectId,
+    }
+
+    let Wrapper { type_, fields } = Wrapper::deserialize(deserializer)?;
+
+    let struct_tag = sui::MoveStructTag::from_str(type_.as_str()).map_err(|e| {
+        serde::de::Error::custom(format!(
+            "Could not parse sui::MoveStructTag from String: {e}"
+        ))
+    })?;
+
+    let Some(tag) = struct_tag.type_params.into_iter().next() else {
+        return Err(serde::de::Error::custom(
+            "Could not get type parameter from `type_`",
+        ));
+    };
+
+    Ok(constructor(tag, fields.id))
+}
+
+fn new_from_fields_id<'de, D, T>(
+    deserializer: D,
+    constructor: impl FnOnce(sui::UID) -> T,
+) -> Result<T, D::Error>
+where
+    D: Deserializer<'de>,
+{
+    #[derive(Deserialize)]
+    struct Wrapper {
+        fields: ObjectId,
+    }
+
+    let Wrapper { fields } = Wrapper::deserialize(deserializer)?;
+
+    Ok(constructor(fields.id))
+}
+
 impl<'de, K, V> Deserialize<'de> for ObjectTable<K, V> {
     fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
     where
         D: Deserializer<'de>,
     {
-        #[derive(Deserialize)]
-        struct Wrapper {
-            #[serde(rename = "type")]
-            type_: String,
-            fields: ObjectId,
-        }
+        Ok(new_from_tag_and_fields_id(deserializer, |tag, id| Self {
+            tag,
+            id,
+            _marker: PhantomData,
+        })?)
+    }
+}
 
-        let Wrapper { type_, fields } = Wrapper::deserialize(deserializer)?;
+/// Move's `LinkedTable` wrapper. This wraps objects that are not stored
+/// directly on the fetched object - one must fetch them separately.
+///
+/// If one value is fetched based on its name then the result is `Response<V>`
+/// where `V` is the generic of `LinkedTable`.
+///
+/// If multiple values are fetched, then the result is `HashMap<K, V>`.
+#[derive(Clone, Debug)]
+pub struct LinkedTable<K, V> {
+    /// The ID of the dynamic object.
+    id: sui::UID,
+    /// Fetching an LinkedTable automatically gives us the key type.
+    tag: sui::MoveTypeTag,
+    _marker: PhantomData<(K, V)>,
+}
 
-        let struct_tag = sui::MoveStructTag::from_str(type_.as_str()).map_err(|e| {
-            serde::de::Error::custom(format!(
-                "Could not parse sui::MoveStructTag from String: {e}"
-            ))
-        })?;
-
-        let Some(tag) = struct_tag.type_params.into_iter().next() else {
-            return Err(serde::de::Error::custom(
-                "Could not get type parameter from `type_`",
-            ));
+impl<K, V> LinkedTable<K, V>
+where
+    K: Eq + Hash + DeserializeOwned + Serialize,
+    V: DeserializeOwned,
+{
+    /// Fetch a single object from the linked table by its key.
+    pub async fn fetch_one(&self, sui: &sui::Client, key: K) -> anyhow::Result<V> {
+        let field_name = sui::DynamicFieldName {
+            type_: self.tag.clone(),
+            value: serde_json::to_value(key).map_err(anyhow::Error::new)?,
         };
 
-        Ok(Self {
+        Ok(
+            dynamic_fetch_one::<ObjectFields<ObjectValue<ObjectFields<ObjectValue<V>>>>>(
+                sui,
+                *self.id.object_id(),
+                field_name,
+            )
+            .await?
+            .fields
+            .value
+            .fields
+            .value,
+        )
+    }
+
+    /// Fetch all objects from the linked table.
+    pub async fn fetch_all(&self, sui: &sui::Client) -> anyhow::Result<HashMap<K, V>> {
+        let response = dynamic_fetch_many::<
+            K,
+            ObjectFields<ObjectValue<ObjectFields<ObjectValue<V>>>>,
+        >(sui, *self.id.object_id())
+        .await?;
+
+        Ok(response
+            .into_iter()
+            .map(|(key, ObjectFields { fields })| (key, fields.value.fields.value))
+            .collect())
+    }
+}
+
+impl<'de, K, V> Deserialize<'de> for LinkedTable<K, V> {
+    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
+    where
+        D: Deserializer<'de>,
+    {
+        Ok(new_from_tag_and_fields_id(deserializer, |tag, id| Self {
             tag,
-            id: fields.id,
+            id,
             _marker: PhantomData,
-        })
+        })?)
     }
 }
 
@@ -147,32 +244,11 @@ impl<'de, K, V> Deserialize<'de> for Table<K, V> {
     where
         D: Deserializer<'de>,
     {
-        #[derive(Deserialize)]
-        struct Wrapper {
-            #[serde(rename = "type")]
-            type_: String,
-            fields: ObjectId,
-        }
-
-        let Wrapper { type_, fields } = Wrapper::deserialize(deserializer)?;
-
-        let struct_tag = sui::MoveStructTag::from_str(type_.as_str()).map_err(|e| {
-            serde::de::Error::custom(format!(
-                "Could not parse sui::MoveStructTag from String: {e}"
-            ))
-        })?;
-
-        let Some(tag) = struct_tag.type_params.into_iter().next() else {
-            return Err(serde::de::Error::custom(
-                "Could not get type parameter from `type_`",
-            ));
-        };
-
-        Ok(Self {
+        Ok(new_from_tag_and_fields_id(deserializer, |tag, id| Self {
             tag,
-            id: fields.id,
+            id,
             _marker: PhantomData,
-        })
+        })?)
     }
 }
 
@@ -228,17 +304,10 @@ impl<'de, K, V> Deserialize<'de> for ObjectBag<K, V> {
     where
         D: Deserializer<'de>,
     {
-        #[derive(Deserialize)]
-        struct Wrapper {
-            fields: ObjectId,
-        }
-
-        let Wrapper { fields } = Wrapper::deserialize(deserializer)?;
-
-        Ok(Self {
-            id: fields.id,
+        Ok(new_from_fields_id(deserializer, |id| Self {
+            id,
             _marker: PhantomData,
-        })
+        })?)
     }
 }
 
@@ -311,17 +380,10 @@ impl<'de, K, V> Deserialize<'de> for Bag<K, V> {
     where
         D: Deserializer<'de>,
     {
-        #[derive(Deserialize)]
-        struct Wrapper {
-            fields: ObjectId,
-        }
-
-        let Wrapper { fields } = Wrapper::deserialize(deserializer)?;
-
-        Ok(Self {
-            id: fields.id,
+        Ok(new_from_fields_id(deserializer, |id| Self {
+            id,
             _marker: PhantomData,
-        })
+        })?)
     }
 }