+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [`0.2.0`] - Unreleased
+## [`0.2.0`] - 2025-08-12
 
 ### Repository
 
@@ -13,7 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - CONTRIBUTING.md
 - CODE_OF_CONDUCT.md
-- `pre-commit` hook
+- `pre-commit` hook (also in CI)
 
 ### `nexus-cli`
 
@@ -39,6 +39,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - configured `cargo-deny` rules
 - not failing if a tool is already registered when registering a tool
 - not failing a whole tool registration batch if one of the tools fails to register
+- `nexus gas limited-invocations enable` to enable the limited invocations gas extension for a tool
+- `nexus gas limited-invocations disable` to disable the limited invocations gas extension for a tool
+- `nexus gas limited-invocations buy-ticket` to buy a limited invocations gas ticket for a tool
 - `--no-save` flag to `nexus tool register` to not save the owner caps to the CLI config
 
 #### Changed
@@ -85,6 +88,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `pre_key_vault` key to `NexusObjects`
 - pre key vault related Nexus events and their definitions
 - DAG validation (moved from `nexus-cli`)
+- `LinkedTable` support for object crawler
+- added identifiers to `tool_registry`'s allow list functions
 
 #### Changed